requirements.txt
pypdf==4.2.0
regex==2024.4.16
torch==2.2.2
tqdm==4.66.2
wandb==0.16.6
